Update Gui Tweaks Mod
Change Gui Blur Mod to Gui Tweaks Mod
Added Option to disable vanilla background
