+++ 1.12.2/cloudclient/src/main/resources/assets/cloudmc/icon/button/button/gui tweaks.png
