+++ 1.12.2/cloudclient/src/main/resources/mixins.cloudmc.json
@@ -9,6 +9,7 @@
     "GuiIngameMenuMixin",
     "GuiInGameMixin",
     "GuiMainMenuMixin",
+    "GuiScreenMixin",
     "ItemRendererMixin",
     "MinecraftMixin"
   ],
