+++ 1.7.10/cloudclient/src/main/java/dev/cloudmc/mixins/GuiScreenMixin.java
@@ -20,6 +20,7 @@ import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
@@ -78,5 +79,18 @@ public abstract class GuiScreenMixin extends Gui {
         }
 
         return scroll;
+    }
+
+    @Redirect(
+            method = "drawWorldBackground",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/client/gui/GuiScreen;drawGradientRect(IIIIII)V"
+            )
+    )
+    public void drawWorldBackground(GuiScreen instance, int left, int top, int right, int bottom, int startColor, int endColor) {
+        if (Cloud.INSTANCE.settingManager.getSettingByModAndName("Gui Tweaks", "Darken Background").isCheckToggled()) {
+            this.drawGradientRect(left, top, right, bottom, startColor, endColor);
+        }
     }
 }
