+++ 1.7.10/cloudclient/src/main/resources/assets/cloudmc/icon/button/button/gui tweaks.png
