+++ 1.8.9/cloudclient/src/main/java/dev/cloudmc/feature/mod/ModManager.java
@@ -41,7 +41,7 @@ public class ModManager {
         addMod(new FreelookMod());
         addMod(new CrosshairMod());
         addMod(new MotionblurMod());
-        addMod(new GuiBlurMod());
+        addMod(new GuiTweaksMod());
         addMod(new BlockOverlayMod());
         addMod(new BlockInfoMod());
         addMod(new ReachDisplayMod());
