+++ 1.8.9/cloudclient/src/main/java/dev/cloudmc/feature/mod/impl/GuiTweaksMod.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2022 DupliCAT
+ * GNU Lesser General Public License v3.0
+ */
+
+package dev.cloudmc.feature.mod.impl;
+
+import dev.cloudmc.Cloud;
+import dev.cloudmc.feature.mod.Mod;
+import dev.cloudmc.feature.mod.Type;
+import dev.cloudmc.feature.setting.Setting;
+import net.minecraft.client.gui.GuiChat;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.client.event.GuiOpenEvent;
+import net.minecraftforge.client.event.GuiScreenEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+
+public class GuiTweaksMod extends Mod {
+
+    public GuiTweaksMod() {
+        super(
+                "Gui Tweaks",
+                "Adds Tweaks to the Gui like blur and transparency.",
+                Type.Tweaks
+        );
+
+        Cloud.INSTANCE.settingManager.addSetting(new Setting("Blur Background", this, true));
+        Cloud.INSTANCE.settingManager.addSetting(new Setting("Darken Background", this, true));
+    }
+
+    @SubscribeEvent
+    public void onGuiOpen(GuiOpenEvent e) {
+        if (Cloud.INSTANCE.settingManager.getSettingByModAndName(getName(), "Blur Background").isCheckToggled()) {
+            if (!(e.gui instanceof GuiChat)) {
+                try {
+                    Cloud.INSTANCE.mc.entityRenderer.loadShader(
+                            new ResourceLocation("shaders/post/blur.json"));
+                } catch (Exception exception) {
+                    System.out.println(exception.getMessage());
+                }
+            }
+            if (e.gui == null) {
+                if (Cloud.INSTANCE.mc.entityRenderer.getShaderGroup() != null) {
+                    Cloud.INSTANCE.mc.entityRenderer.getShaderGroup().deleteShaderGroup();
+                }
+            }
+        }
+    }
+}
